moves files for the time being
